+++ src/view/content-details-inner-view.js
